# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/user/{id}": {
      "get": {
        "operationId": "UserController_getById",
        "summary": "",
        "description": "Get user by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "summary": "",
        "description": "Update user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "delete": {
        "operationId": "UserController_delete",
        "summary": "",
        "description": "Delete user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user": {
      "post": {
        "operationId": "UserController_getUser",
        "summary": "",
        "description": "Get all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDocument"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
    "/scoring/{id}": {
      "get": {
        "operationId": "ScoringController_getById",
        "summary": "",
        "description": "Get scoring by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringDTO"
                }
              }
            }
          }
        },
        "tags": [
          "scoring"
        ]
      },
      "patch": {
        "operationId": "ScoringController_update",
        "summary": "",
        "description": "Update scoring",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoringUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringDTO"
                }
              }
            }
          }
        },
        "tags": [
          "scoring"
        ]
      },
      "delete": {
        "operationId": "ScoringController_delete",
        "summary": "",
        "description": "Delete scoring",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringDTO"
                }
              }
            }
          }
        },
        "tags": [
          "scoring"
        ]
      }
    },
    "/scoring": {
      "get": {
        "operationId": "ScoringController_getAll",
        "summary": "",
        "description": "Get all scorings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringDTO"
                }
              }
            }
          }
        },
        "tags": [
          "scoring"
        ]
      },
      "post": {
        "operationId": "ScoringController_create",
        "summary": "",
        "description": "Create new scoring",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoringCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringDTO"
                }
              }
            }
          }
        },
        "tags": [
          "scoring"
        ]
      }
    },
...
  },
  "info": {
    "title": "BlockTrust API",
    "description": "The BlockTrust API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "UserDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "aboutMe": {
            "type": "string"
          },
          "orcid": {
            "type": "string"
          },
          "linkedin": {
            "type": "string"
          },
          "blog": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "starredPapers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isOnboarded": {
            "type": "boolean"
          },
          "acceptedTC": {
            "type": "boolean"
          },
          "disciplines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "userType": {
            "enum": [
              "student",
              "medical",
              "business",
              "researcher",
              "citizen",
              "academic"
            ],
            "type": "string"
          },
          "sub": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "firstName",
          "lastName",
          "email",
          "aboutMe",
          "orcid",
          "linkedin",
          "blog",
          "role",
          "starredPapers",
          "isOnboarded",
          "acceptedTC",
          "disciplines",
          "userType",
          "sub"
        ]
      },
      "UserDocument": {
        "type": "object",
        "properties": {}
      },
      "UserUpdateDTO": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "aboutMe": {
            "type": "string"
          },
          "orcid": {
            "type": "string"
          },
          "linkedin": {
            "type": "string"
          },
          "blog": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "starredPapers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isOnboarded": {
            "type": "boolean"
          },
          "acceptedTC": {
            "type": "boolean"
          },
          "disciplines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "userType": {
            "type": "string",
            "enum": [
              "student",
              "medical",
              "business",
              "researcher",
              "citizen",
              "academic"
            ]
          }
        }
      },
...
      "ScoringDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "reputationModel": {
            "type": "string"
          },
          "paper": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "_id",
          "reputationModel",
          "paper",
          "score",
          "timestamp"
        ]
      },
...
      "ScoringCreateDTO": {
        "type": "object",
        "properties": {
          "reputationModel": {
            "type": "string"
          },
          "paper": {
            "type": "string"
          }
        },
        "required": [
          "reputationModel",
          "paper"
        ]
      },
      "ScoringUpdateDTO": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "score",
          "timestamp"
        ]
      },
...
    }
  }
}